melhorias gerais do caso de uso gerenciar aula

+++ spondereServer/controller/academicClass.py
@@ -1,30 +1,55 @@
 from fastapi import (
     APIRouter, 
-    Depends)
-from controller.security import (
-    JWTBearer)
+    Depends, 
+    Request)
+from controller.security import JWTBearer, getCurrentUserName
 from entity.academicClass import AcademicClass
-from database import academicClass as classDB, discipline
+from database import academicClass as classDB
+from database import user as userDB
+from fastapi import HTTPException
 
 router = APIRouter()
  
-
 @router.post("", dependencies=[Depends(JWTBearer())])
-async def createAcademicClass(academicClass:AcademicClass) -> dict:
+async def createAcademicClass(academicClass:AcademicClass, request:Request) -> dict:
+    authorization = request.headers.get("authorization")
+    userName = getCurrentUserName(authorization)
+    user = userDB.read(userName)
+
+    if not user.administrator or not user.professor:
+        raise HTTPException(status_code=401,
+            detail="O usuario não tem privilegio de administrador ou professor.")
+    
     id = classDB.create(academicClass)
     return {
         "id": id
     }
 
 @router.put("/{id}", dependencies=[Depends(JWTBearer())])
-async def updateAcademicClass(id:int, academicClass:AcademicClass)-> dict:
+async def updateAcademicClass(id:int, academicClass:AcademicClass, request:Request)-> dict:
+    authorization = request.headers.get("authorization")
+    userName = getCurrentUserName(authorization)
+    user = userDB.read(userName)
+
+    if not user.administrator or not user.professor:
+        raise HTTPException(status_code=401,
+            detail="O usuario não tem privilegio de administrador ou professor.")
+
     classDB.update(id, academicClass)
     return{
         "result": "success"
     }
 
 @router.delete("/{id}", dependencies=[Depends(JWTBearer())])
-async def deleteAcademicClass(id:int) -> dict:
+async def deleteAcademicClass(id:int, request:Request) -> dict:
+    authorization = request.headers.get("authorization")
+    userName = getCurrentUserName(authorization)
+    user = userDB.read(userName)
+
+    if not user.administrator or not user.professor:
+        raise HTTPException(status_code=401,
+            detail="O usuario não tem privilegio de administrador ou professor.")
+
     classDB.delete(id)
     return{
          "result": "success"
